requirements_cpu.txt
absl-py==0.7.1
astor==0.8.0
bleach==1.5.0
gast==0.2.2
grpcio==1.20.1
html5lib==0.9999999
joblib==0.13.2
Markdown==3.1.1
numpy==1.16.4
opencv-python==3.4.3.18
pandas==0.24.1
Pillow==5.4.1
protobuf==3.6.0
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.21.1
scipy==1.3.0
six==1.12.0
sklearn==0.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
Werkzeug==0.15.4
